Fix DeBERTa fp16/fp32 dtype mismatch: cast logits to float32 before loss
Error: RuntimeError: expected scalar type Half but found Float. Root cause: DeBERTa-v3-base loads its weights in fp16 on GPU by default (even without AMP). The CrossEntropyLoss class weight tensor was fp32. PyTorch does not auto-cast criterion inputs. Fix: outputs.logits.float() before passing to criterion — ensures the loss computation always runs in fp32 regardless of the model's internal dtype. This is safe for all models (FinBERT, DistilBERT, DeBERTa). Notebook v6: sanity check now also verifies the float cast. 115 tests pass.

+++ src/model.py
@@ -176,7 +176,9 @@ class FinancialSentimentModel:
 
                 with torch.autocast(device_type="cuda" if use_amp else "cpu", enabled=use_amp):
                     outputs = self.model(input_ids, attention_mask=attention_mask)
-                    loss = criterion(outputs.logits, labels)
+                    # Cast logits to float32 for the loss — required when the model
+                    # runs in fp16 (e.g. DeBERTa-v3) but the criterion weight is fp32.
+                    loss = criterion(outputs.logits.float(), labels)
                 train_loss += loss.item()
 
                 scaler.scale(loss).backward()
